it's been a while, idk what the last changes I made were

+++ src/fiber.h
@@ -45,7 +45,7 @@ public:
 private:
 
 	void clear();
-	bool isGoodSingle(Event&, double, double);
+	bool isGoodSingle(Event&, int, int/*, double, double*/);
 };
 
 #endif
